#71 removing jsonification in Read Describe results
