+++ occurrence-ws/src/test/java/org/gbif/occurrence/ws/provider/hive/SQLDescribeAndExplainTest.java
@@ -1,5 +1,8 @@
 package org.gbif.occurrence.ws.provider.hive;
 
+import java.io.IOException;
+import java.util.List;
+import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
@@ -7,32 +10,33 @@ import org.junit.rules.ExpectedException;
 
 
 public class SQLDescribeAndExplainTest {
-
+  private static final ObjectMapper mapper = new ObjectMapper();
   @Rule
   public ExpectedException expectedEx = ExpectedException.none();
   
   @Test
   public void testLegalExplain() {
-    String result = new HiveSQL.Execute().explain("SELECT * FROM occurrence_hdfs");
+    String result = HiveSQL.Execute.explain("SELECT * FROM occurrence_hdfs");
     Assert.assertTrue(!result.isEmpty());
   }
 
   @Test
   public void testIllLegalExplain() {
     expectedEx.expect(RuntimeException.class);
-    String result = new HiveSQL.Execute().explain("SELECT * FROM occurre");
+    String result = HiveSQL.Execute.explain("SELECT * FROM occurre");
     System.out.println(result);
   }
   
   @Test
-  public void testValidDescribe()  {
-    String result = new HiveSQL.Execute().describe("occurrence_hdfs");
+  public void testValidDescribe() throws IOException  {
+    List<DescribeResult> result = HiveSQL.Execute.describe("occurrence_hdfs");
     Assert.assertTrue(!result.isEmpty());
+    System.out.println(mapper.writeValueAsString(result));
   }
   
   @Test
   public void testInValidDescribe()  {
     expectedEx.expect(RuntimeException.class);
-    new HiveSQL.Execute().describe("occurrence");
+    HiveSQL.Execute.describe("occurrence");
   }
 }
